requirements.txt
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.30.1
Flask==3.0.3
Flask-Cors==4.0.0
numpy==1.26.4
opencv-python-headless==4.9.0.80